Refactor search and school maps to use BaseMap
Probably a better refactoring can be done once the search
map get's more features.
